Update Spinner, de instellingen worden nu opgehaald per gebruiker (op basis van gebruikersnaam)

+++ java/src/MainInput.java
@@ -252,7 +252,7 @@ public class MainInput {
 
             double TempVerwarmen =0.0;
             while (rs.next()) {
-                TempVerwarmen = rs.getInt(1);
+                TempVerwarmen = rs.getDouble(1);
             }
             conn.close();
